// src/utils/Enums.ts
export enum THEME_TYPE {
    Light = 'light',
    Dark = 'dark',
  }

  export enum BOTTOM_TAB_NAVIGATOR_SCREENS {
    Home = 'Home',
    Booking = 'Booking',
    Search = 'Search',
    WishList = 'WishList',
    Account = 'Account'
  }

  export enum STACK_SCREENS {
    Home = 'HomeScreen',
    EventDetails = 'EventDetails',
    SeatSelactions = 'SeatSelection',
    Payments = 'Payments'
  }

  export enum SVG_ICONS {
    HomeIcon = 'HomeIcon',
    HappningLogo = 'HappningLogo',
    SocialMediaIcon  = 'SocialMediaIcon',
    MenuIcon = 'MenuIcon',
    NotificationIcon = 'NotificationIcon',
    LocationIcon = 'LocationIcon',
    AccountIcon = 'AccountIcon',
    SearchIcon = 'SearchIcon',
    WishlistIcon = 'WishlistIcon',
    BookingIcon = 'BookingIcon',
    BackIcon = 'BackIcon',
    ShareIcon = 'ShareIcon'
  }

  export enum TEXT_TYPE {
    PageTitle = 'pageTitle', //18 , 500
    SectionTitle = 'sectionTitle', //16 ,500
    PrimaryText = 'primaryText', //14 , 400
    SecondaryText = 'secondaryText', //12 , 400
    ErrorText = 'errorText', //12 ,400
    TextDisabled = 'textDisabled'
  }